[llz] 接口请求过滤

+++ src/main/java/com/softlab/okr/mapper/SwitchMapper.java
@@ -1,9 +1,10 @@
 package com.softlab.okr.mapper;
 
 import com.softlab.okr.model.entity.Switch;
-import java.util.List;
 import org.springframework.dao.DataAccessException;
 import org.springframework.stereotype.Repository;
+
+import java.util.List;
 
 @Repository
 public interface SwitchMapper {
@@ -12,7 +13,7 @@ public interface SwitchMapper {
 
   int updateStatus(String name) throws DataAccessException;
 
-  Byte selectStatus(String name) throws DataAccessException;
+  Integer selectStatus(String name) throws DataAccessException;
 
   int deleteByName(String name) throws DataAccessException;
