+++ src/main/java/com/softlab/okr/service/SwitchService.java
@@ -2,6 +2,7 @@ package com.softlab.okr.service;
 
 import com.softlab.okr.exception.ServiceException;
 import com.softlab.okr.model.entity.Switch;
+
 import java.util.List;
 
 public interface SwitchService {
@@ -10,7 +11,7 @@ public interface SwitchService {
 
   int modifyStatus(String name) throws ServiceException;
 
-  Byte getStatus(String name) throws ServiceException;
+  Integer getStatus(String name) throws ServiceException;
 
   int removeByName(String name) throws ServiceException;
